Remove unnecessary cast

+++ inc/libalx/base/stdlib/popcnt.h
@@ -103,8 +103,8 @@ uint64_t	popcnt_u64	(uint64_t n)
 inline
 uint64_t	popcnt_u128	(uint128_t n)
 {
-	const	uint_fast64_t	cnt_a	= popcnt_u64((uint64_t)n);
-	const	uint_fast64_t	cnt_b	= popcnt_u64((uint64_t)(n >> 64));
+	const	uint_fast64_t	cnt_a	= popcnt_u64(n);
+	const	uint_fast64_t	cnt_b	= popcnt_u64(n >> 64);
 	const	uint_fast64_t	cnt	= cnt_a + cnt_b;
 
 	return	cnt;
@@ -115,10 +115,10 @@ uint64_t	popcnt_u128	(uint128_t n)
 inline
 uint64_t	popcnt_u256	(uint256_t n)
 {
-	const	uint_fast64_t	cnt_a	= popcnt_u64((uint64_t)n);
-	const	uint_fast64_t	cnt_b	= popcnt_u64((uint64_t)(n >> 64));
-	const	uint_fast64_t	cnt_c	= popcnt_u64((uint64_t)(n >> (64 * 2)));
-	const	uint_fast64_t	cnt_d	= popcnt_u64((uint64_t)(n >> (64 * 3)));
+	const	uint_fast64_t	cnt_a	= popcnt_u64(n);
+	const	uint_fast64_t	cnt_b	= popcnt_u64(n >> 64);
+	const	uint_fast64_t	cnt_c	= popcnt_u64(n >> (64 * 2));
+	const	uint_fast64_t	cnt_d	= popcnt_u64(n >> (64 * 3));
 	const	uint_fast64_t	cnt	= cnt_a + cnt_b + cnt_c + cnt_d;
 
 	return	cnt;
